move editable input
